requirements.txt
streamlit>=1.28.0
sounddevice>=0.4.6
scipy>=1.10.0
librosa>=0.10.0
numpy>=1.21.0
tensorflow-cpu>=2.10.0
tensorflow-hub>=0.12.0
matplotlib>=3.5.0
pydub>=0.25.0
pandas>=1.5.0
soundfile>=0.12.0